+++ lldb/source/Plugins/Language/ObjC/NSError.cpp
@@ -124,7 +124,7 @@ public:
     return 0;
   }
 
-  lldb::ValueObjectSP GetChildAtIndex(size_t idx) override {
+  lldb::ValueObjectSP GetChildAtIndex(uint64_t idx) override {
     if (idx != 0)
       return lldb::ValueObjectSP();
